pyproject.toml
[project]
name = "agentic-architect"
version = "0.1.0"
description = "AgenticArchitect - All as Code & AI Agents"
authors = [
    { name = "fabienfrfr", email = "fabien.furfaro@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "architect",
]

[tool.uv]
managed = true
default-groups = ["dev"]

[tool.uv.sources]
architect = { workspace = true }

[tool.uv.workspace]
members = ["apps/architect"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "jupytext>=1.19.1",
    "marimo>=0.19.6",
    "matplotlib>=3.10.8",
    "nest-asyncio>=1.6.0",
    "pathspec>=1.0.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-bdd>=8.1.0",
    "pytest-dotenv>=0.5.2",
    "pytest-env>=1.2.0",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

env_files = [".env"]
env = [
    "ENV=test",
    "PYTHONPATH=.",
    "PHOENIX_PORT=6006",
    "PHOENIX_COLLECTOR_ENDPOINT=http://localhost:4317",
    "OLLAMA_URL=http://localhost:11434"
]


markers = [
    "infra: Infrastructure connectivity tests",
    "bdd: Behavioral Driven Development tests",
    "integration: Tests requiring the full cluster"
]
